requirements.txt
pandas
python-dotenv
requests
psycopg2-binary
SQLAlchemy
